Grid/table generation, with click event
Added the table generation with column and row ids.
Added a click event to turn on the cells for the initial start
Instead of creating a long list of if statements (cell alive, cell
dead), I broke up the if statements into seperate functions for every
event of the alive/dead cell.